Create openDICOM component
